# Improve query-scheduler performance under load