--- a/pkg/scheduler/queue/queue.go
+++ b/pkg/scheduler/queue/queue.go
@@ -7,7 +7,7 @@ package queue
 
 import (
 	"context"
-	"sync"
+	"fmt"
 	"time"
 
 	"github.com/grafana/dskit/services"
@@ -54,29 +54,235 @@ type Request interface{}
 type RequestQueue struct {
 	services.Service
 
+	maxOutstandingPerTenant int
+	forgetDelay             time.Duration
+
 	connectedQuerierWorkers *atomic.Int32
 
-	mtx     sync.Mutex
-	cond    contextCond // Notified when request is enqueued or dequeued, or querier is disconnected.
-	queues  *queues
-	stopped bool
+	stopRequested     chan struct{} // Closed by stop() to wake up dispatcherLoop() in response to a stop request.
+	stopCompleted     chan struct{} // Closed by dispatcherLoop() after a stop is requested and the dispatcher has stopped.
+	querierOperations chan querierOperation
+	enqueueRequests   chan enqueueRequest
+	availableQueriers chan *availableQuerier
 
 	queueLength       *prometheus.GaugeVec   // Per user and reason.
 	discardedRequests *prometheus.CounterVec // Per user.
 }
 
+type querierOperation struct {
+	querierID string
+	operation querierOperationType
+	processed chan struct{}
+}
+
+type querierOperationType int
+
+const (
+	registerConnection querierOperationType = iota
+	unregisterConnection
+	notifyShutdown
+	forgetDisconnected
+)
+
+type enqueueRequest struct {
+	userID      string
+	req         Request
+	maxQueriers int
+	successFn   func()
+	processed   chan error
+}
+
+type availableQuerier struct {
+	ctx           context.Context
+	querierID     string
+	lastUserIndex UserIndex
+	processed     chan nextRequestForQuerier
+	next          *availableQuerier
+}
+
+type nextRequestForQuerier struct {
+	req  Request
+	next UserIndex
+	err  error
+}
+
 func NewRequestQueue(maxOutstandingPerTenant int, forgetDelay time.Duration, queueLength *prometheus.GaugeVec, discardedRequests *prometheus.CounterVec) *RequestQueue {
 	q := &RequestQueue{
-		queues:                  newUserQueues(maxOutstandingPerTenant, forgetDelay),
+		maxOutstandingPerTenant: maxOutstandingPerTenant,
+		forgetDelay:             forgetDelay,
 		connectedQuerierWorkers: atomic.NewInt32(0),
 		queueLength:             queueLength,
 		discardedRequests:       discardedRequests,
 	}
 
-	q.cond = contextCond{Cond: sync.NewCond(&q.mtx)}
-	q.Service = services.NewTimerService(forgetCheckPeriod, nil, q.forgetDisconnectedQueriers, q.stopping).WithName("request queue")
+	q.Service = services.NewTimerService(forgetCheckPeriod, q.starting, q.forgetDisconnectedQueriers, q.stop).WithName("request queue")
 
 	return q
+}
+
+func (q *RequestQueue) starting(_ context.Context) error {
+	q.stopRequested = make(chan struct{})
+	q.stopCompleted = make(chan struct{})
+
+	// These channels must not be buffered so that we can detect when dispatcherLoop() has finished.
+	q.querierOperations = make(chan querierOperation)
+	q.enqueueRequests = make(chan enqueueRequest)
+	q.availableQueriers = make(chan *availableQuerier)
+
+	go q.dispatcherLoop()
+
+	return nil
+}
+
+func (q *RequestQueue) dispatcherLoop() {
+	stopping := false
+	queues := newUserQueues(q.maxOutstandingPerTenant, q.forgetDelay)
+
+	var firstAvailableQuerier *availableQuerier
+	var lastAvailableQuerier *availableQuerier
+
+	for {
+		needToDispatchQueries := false
+
+		select {
+		case <-q.stopRequested:
+			// Nothing much to do here - fall through to the stop logic below to see if we can stop immediately.
+			stopping = true
+		case qe := <-q.querierOperations:
+			switch qe.operation {
+			case registerConnection:
+				queues.addQuerierConnection(qe.querierID)
+			case unregisterConnection:
+				queues.removeQuerierConnection(qe.querierID, time.Now())
+			case notifyShutdown:
+				queues.notifyQuerierShutdown(qe.querierID)
+			case forgetDisconnected:
+				if queues.forgetDisconnectedQueriers(time.Now()) > 0 {
+					// Removing some queriers may have caused a resharding.
+					needToDispatchQueries = true
+				}
+			default:
+				panic(fmt.Sprintf("received unknown querier event %v for querier ID %v", qe.operation, qe.querierID))
+			}
+			qe.processed <- struct{}{}
+		case r := <-q.enqueueRequests:
+			err := q.handleNewRequest(queues, r)
+			r.processed <- err
+
+			if err == nil {
+				// TODO: might be able to be much smarter here and try to find a querier that can take the request directly
+				needToDispatchQueries = true
+			}
+		case querier := <-q.availableQueriers:
+			if !q.dispatchRequestToQuerier(queues, querier) {
+				// No requests available for this querier right now. Add it to the list to try later.
+				if lastAvailableQuerier == nil {
+					firstAvailableQuerier = querier
+					lastAvailableQuerier = querier
+				} else {
+					lastAvailableQuerier.next = querier
+					lastAvailableQuerier = querier
+				}
+			}
+		}
+
+		if needToDispatchQueries {
+			var previousQuerier *availableQuerier
+			currentQuerier := firstAvailableQuerier
+
+			for currentQuerier != nil && queues.len() > 0 {
+				if q.dispatchRequestToQuerier(queues, currentQuerier) {
+					// Remove this querier from our list.
+
+					if previousQuerier == nil {
+						// This was the first available querier.
+						firstAvailableQuerier = currentQuerier.next
+					} else {
+						previousQuerier.next = currentQuerier.next
+					}
+
+					if currentQuerier.next == nil {
+						// This was the last available querier.
+						lastAvailableQuerier = previousQuerier
+					}
+
+					currentQuerier = currentQuerier.next
+				} else {
+					// Nothing for this querier right now, advance through the list.
+					previousQuerier = currentQuerier
+					currentQuerier = currentQuerier.next
+				}
+			}
+		}
+
+		// TODO: old logic stopped when queue was empty or there were no workers, but I'm not sure that's quite right - should we wait for both?
+		if stopping && (queues.len() == 0 || q.connectedQuerierWorkers.Load() == 0) {
+			// Tell any waiting GetNextRequestForQuerier calls that nothing is coming.
+			for firstAvailableQuerier != nil {
+				firstAvailableQuerier.processed <- nextRequestForQuerier{err: ErrStopped}
+				firstAvailableQuerier = firstAvailableQuerier.next
+			}
+
+			// We are done.
+			close(q.stopCompleted)
+			return
+		}
+	}
+}
+
+func (q *RequestQueue) handleNewRequest(queues *queues, r enqueueRequest) error {
+	queue := queues.getOrAddQueue(r.userID, r.maxQueriers)
+	if queue == nil {
+		// This can only happen if userID is "".
+		return errors.New("no queue found")
+	}
+
+	select {
+	case queue <- r.req:
+		q.queueLength.WithLabelValues(r.userID).Inc()
+		// Call the successFn before sending this request to a waiting querier.
+		if r.successFn != nil {
+			r.successFn()
+		}
+		return nil
+	default:
+		q.discardedRequests.WithLabelValues(r.userID).Inc()
+		return ErrTooManyRequests
+	}
+}
+
+// dispatchRequestToQuerier finds and forwards a request to a querier, if a suitable request is available.
+// Returns true if this querier should be removed from the list of waiting queriers (eg. because a request has been forwarded to it), false otherwise.
+func (q *RequestQueue) dispatchRequestToQuerier(queues *queues, querier *availableQuerier) bool {
+	if querier.ctx.Err() != nil {
+		querier.processed <- nextRequestForQuerier{
+			next: querier.lastUserIndex,
+			err:  querier.ctx.Err(),
+		}
+		return true
+	}
+
+	queue, userID, idx := queues.getNextQueueForQuerier(querier.lastUserIndex.last, querier.querierID)
+	querier.lastUserIndex.last = idx
+	if queue == nil {
+		// Nothing available for this querier, try again next time.
+		return false
+	}
+
+	// Pick next request from the queue.
+	request := <-queue
+	if len(queue) == 0 {
+		queues.deleteQueue(userID)
+	}
+
+	q.queueLength.WithLabelValues(userID).Dec()
+
+	querier.processed <- nextRequestForQuerier{
+		req:  request,
+		next: querier.lastUserIndex,
+		err:  querier.ctx.Err(),
+	}
+	return true
 }
 
 // EnqueueRequest puts the request into the queue. MaxQueries is user-specific value that specifies how many queriers can
@@ -85,31 +291,20 @@ func NewRequestQueue(maxOutstandingPerTenant int, forgetDelay time.Duration, que
 //
 // If request is successfully enqueued, successFn is called with the lock held, before any querier can receive the request.
 func (q *RequestQueue) EnqueueRequest(userID string, req Request, maxQueriers int, successFn func()) error {
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-
-	if q.stopped {
-		return ErrStopped
-	}
-
-	queue := q.queues.getOrAddQueue(userID, maxQueriers)
-	if queue == nil {
-		// This can only happen if userID is "".
-		return errors.New("no queue found")
+	// TODO: pool these?
+	r := enqueueRequest{
+		userID:      userID,
+		req:         req,
+		maxQueriers: maxQueriers,
+		successFn:   successFn,
+		processed:   make(chan error),
 	}
 
 	select {
-	case queue <- req:
-		q.queueLength.WithLabelValues(userID).Inc()
-		q.cond.Broadcast()
-		// Call this function while holding a lock. This guarantees that no querier can fetch the request before function returns.
-		if successFn != nil {
-			successFn()
-		}
-		return nil
-	default:
-		q.discardedRequests.WithLabelValues(userID).Inc()
-		return ErrTooManyRequests
+	case q.enqueueRequests <- r:
+		return <-r.processed
+	case <-q.stopCompleted:
+		return ErrStopped
 	}
 }
 
@@ -117,81 +312,42 @@ func (q *RequestQueue) EnqueueRequest(userID string, req Request, maxQueriers in
 // By passing user index from previous call of this method, querier guarantees that it iterates over all users fairly.
 // If querier finds that request from the user is already expired, it can get a request for the same user by using UserIndex.ReuseLastUser.
 func (q *RequestQueue) GetNextRequestForQuerier(ctx context.Context, last UserIndex, querierID string) (Request, UserIndex, error) {
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-
-	querierWait := false
-
-FindQueue:
-	// We need to wait if there are no users, or no pending requests for given querier.
-	for (q.queues.len() == 0 || querierWait) && ctx.Err() == nil && !q.stopped {
-		querierWait = false
-		q.cond.Wait(ctx)
+	// TODO: pool these?
+	querier := &availableQuerier{
+		ctx:           ctx,
+		querierID:     querierID,
+		lastUserIndex: last,
+		processed:     make(chan nextRequestForQuerier),
 	}
 
-	if q.stopped {
+	select {
+	case q.availableQueriers <- querier:
+		result := <-querier.processed
+		return result.req, result.next, result.err
+	case <-q.stopCompleted:
 		return nil, last, ErrStopped
 	}
-
-	if err := ctx.Err(); err != nil {
-		return nil, last, err
-	}
-
-	for {
-		queue, userID, idx := q.queues.getNextQueueForQuerier(last.last, querierID)
-		last.last = idx
-		if queue == nil {
-			break
-		}
-
-		// Pick next request from the queue.
-		for {
-			request := <-queue
-			if len(queue) == 0 {
-				q.queues.deleteQueue(userID)
-			}
-
-			q.queueLength.WithLabelValues(userID).Dec()
-
-			// Tell stopping() we've processed a request.
-			q.cond.Broadcast()
-
-			return request, last, nil
-		}
-	}
-
-	// There are no unexpired requests, so we can get back
-	// and wait for more requests.
-	querierWait = true
-	goto FindQueue
 }
 
 func (q *RequestQueue) forgetDisconnectedQueriers(_ context.Context) error {
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
+	op := querierOperation{
+		operation: forgetDisconnected,
+		processed: make(chan struct{}),
+	}
 
-	if q.queues.forgetDisconnectedQueriers(time.Now()) > 0 {
-		// We need to notify goroutines cause having removed some queriers
-		// may have caused a resharding.
-		q.cond.Broadcast()
+	select {
+	case q.querierOperations <- op:
+		<-op.processed
+	case <-q.stopCompleted:
+		// Nothing to do.
 	}
 
 	return nil
 }
 
-func (q *RequestQueue) stopping(_ error) error {
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-
-	for q.queues.len() > 0 && q.connectedQuerierWorkers.Load() > 0 {
-		q.cond.Wait(context.Background())
-	}
-
-	// Only stop after dispatching enqueued requests.
-	q.stopped = true
-
-	// If there are still goroutines in GetNextRequestForQuerier method, they get notified.
-	q.cond.Broadcast()
+func (q *RequestQueue) stop(_ error) error {
+	q.stopRequested <- struct{}{} // Why not close the channel? We only want to trigger dispatcherLoop() once.
+	<-q.stopCompleted
 
 	return nil
 }
@@ -199,84 +355,54 @@ func (q *RequestQueue) stopping(_ error) error {
 func (q *RequestQueue) RegisterQuerierConnection(querierID string) {
 	q.connectedQuerierWorkers.Inc()
 
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-	q.queues.addQuerierConnection(querierID)
+	op := querierOperation{
+		querierID: querierID,
+		operation: registerConnection,
+		processed: make(chan struct{}),
+	}
+
+	select {
+	case q.querierOperations <- op:
+		<-op.processed
+	case <-q.stopCompleted:
+		// TODO: return error?
+	}
 }
 
 func (q *RequestQueue) UnregisterQuerierConnection(querierID string) {
-	q.connectedQuerierWorkers.Dec()
+	q.connectedQuerierWorkers.Inc()
 
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-	q.queues.removeQuerierConnection(querierID, time.Now())
+	op := querierOperation{
+		querierID: querierID,
+		operation: unregisterConnection,
+		processed: make(chan struct{}),
+	}
+
+	select {
+	case q.querierOperations <- op:
+		<-op.processed
+	case <-q.stopCompleted:
+		// TODO: return error?
+	}
 }
 
 func (q *RequestQueue) NotifyQuerierShutdown(querierID string) {
-	q.mtx.Lock()
-	defer q.mtx.Unlock()
-	q.queues.notifyQuerierShutdown(querierID)
+	q.connectedQuerierWorkers.Inc()
+
+	op := querierOperation{
+		querierID: querierID,
+		operation: notifyShutdown,
+		processed: make(chan struct{}),
+	}
+
+	select {
+	case q.querierOperations <- op:
+		<-op.processed
+	case <-q.stopCompleted:
+		// TODO: return error?
+	}
 }
 
 func (q *RequestQueue) GetConnectedQuerierWorkersMetric() float64 {
 	return float64(q.connectedQuerierWorkers.Load())
-}
-
-// contextCond is a *sync.Cond with Wait() method overridden to support context-based waiting.
-type contextCond struct {
-	*sync.Cond
-
-	// testHookBeforeWaiting is called before calling Cond.Wait() if it's not nil.
-	// Yes, it's ugly, but the http package settled jurisprudence:
-	// https://github.com/golang/go/blob/6178d25fc0b28724b1b5aec2b1b74fc06d9294c7/src/net/http/client.go#L596-L601
-	testHookBeforeWaiting func()
-}
-
-// Wait does c.cond.Wait() but will also return if the context provided is done.
-// All the documentation of sync.Cond.Wait() applies, but it's especially important to remember that the mutex of
-// the cond should be held while Wait() is called (and mutex will be held once it returns)
-func (c contextCond) Wait(ctx context.Context) {
-	// "condWait" goroutine does q.cond.Wait() and signals through condWait channel.
-	condWait := make(chan struct{})
-	go func() {
-		if c.testHookBeforeWaiting != nil {
-			c.testHookBeforeWaiting()
-		}
-		c.Cond.Wait()
-		close(condWait)
-	}()
-
-	// "waiting" goroutine: signals that the condWait goroutine has started waiting.
-	// Notice that a closed waiting channel implies that the goroutine above has started waiting
-	// (because it has unlocked the mutex), but the other way is not true:
-	// - condWait it may have unlocked and is waiting, but someone else locked the mutex faster than us:
-	//   in this case that caller will eventually unlock, and we'll be able to enter here.
-	// - condWait called Wait(), unlocked, received a broadcast and locked again faster than we were able to lock here:
-	//   in this case condWait channel will be closed, and this goroutine will be waiting until we unlock.
-	waiting := make(chan struct{})
-	go func() {
-		c.L.Lock()
-		close(waiting)
-		c.L.Unlock()
-	}()
-
-	select {
-	case <-condWait:
-		// We don't know whether the waiting goroutine is done or not, but we don't care:
-		// it will be done once nobody is fighting for the mutex anymore.
-	case <-ctx.Done():
-		// In order to avoid leaking the condWait goroutine, we can send a broadcast.
-		// Before sending the broadcast we need to make sure that condWait goroutine is already waiting (or has already waited).
-		select {
-		case <-condWait:
-			// No need to broadcast as q.cond.Wait() has returned already.
-			return
-		case <-waiting:
-			// q.cond.Wait() might be still waiting (or maybe not!), so we'll poke it just in case.
-			c.Broadcast()
-		}
-
-		// Make sure we are not waiting anymore, we need to do that before returning as the caller will need to unlock the mutex.
-		<-condWait
-	}
 }