--- a/pkg/scheduler/queue/queue.go
+++ b/pkg/scheduler/queue/queue.go
@@ -161,12 +161,13 @@ func (q *RequestQueue) dispatcherLoop() {
 			}
 			qe.processed <- struct{}{}
 		case r := <-q.enqueueRequests:
-			err := q.handleEnqueueRequest(queues, r)
+			firstRequest, err := q.handleEnqueueRequest(queues, r)
 			r.processed <- err
 
 			if err == nil {
-				// TODO: might be able to be much smarter here and try to find a querier that can take the request directly
-				needToDispatchQueries = true
+				// If this isn't the first request for this tenant, we know none of the waiting queriers can process this request, so there's no point checking again:
+				// if any of the waiting queriers could process this request, they would have taken earlier requests in this tenant's queue already.
+				needToDispatchQueries = firstRequest
 			}
 		case querier := <-q.availableQueriers:
 			if !q.dispatchRequestToQuerier(queues, querier) {
@@ -218,24 +219,27 @@ func (q *RequestQueue) dispatcherLoop() {
 	}
 }
 
-func (q *RequestQueue) handleEnqueueRequest(queues *queues, r enqueueRequest) error {
+// handleEnqueueRequest stores a new request in the tenant's queue, and returns true if this was the first request for this tenant,
+// or false otherwise.
+func (q *RequestQueue) handleEnqueueRequest(queues *queues, r enqueueRequest) (bool, error) {
 	queue := queues.getOrAddQueue(r.userID, r.maxQueriers)
 	if queue == nil {
 		// This can only happen if userID is "".
-		return errors.New("no queue found")
+		return false, errors.New("no queue found")
 	}
 
 	select {
 	case queue <- r.req:
+		firstRequest := len(queue) == 1
 		q.queueLength.WithLabelValues(r.userID).Inc()
 		// Call the successFn here to ensure we call it before sending this request to a waiting querier.
 		if r.successFn != nil {
 			r.successFn()
 		}
-		return nil
+		return firstRequest, nil
 	default:
 		q.discardedRequests.WithLabelValues(r.userID).Inc()
-		return ErrTooManyRequests
+		return false, ErrTooManyRequests
 	}
 }