#### What this PR does

This PR contains an alternative implementation of the query-scheduler `RequestQueue` that scales much better with many simultaneous queriers. 

The idea is to use a central dispatcher to allocate requests to queriers, rather than having all queriers compete to take a lock to check if work is available for them every time a new request arrives, and then waste cycles searching for work that has already been taken by another querier.

There's still some work to be done to improve performance further - with the new implementation, latency is far more consistent regardless of the number of queriers, but the new implementation isn't as performant as the previous one for low numbers of queriers. Given the low absolute value of these differences, and the significant improvement for the high concurrency case, I think this is good enough to merge, and I'll continue to iterate on further improvements in subsequent PRs.

<details>
<summary>Benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/scheduler/queue
                                                                                            │ original.txt  │ step-9-simpler-context-cancellation-with-less-memory.txt │
                                                                                            │    sec/op     │              sec/op               vs base                │
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/16_concurrent_consumers-10         543.7n ±  2%                      1756.0n ± 39%  +222.97% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/160_concurrent_consumers-10        939.3n ±  1%                      2133.5n ±  7%  +127.14% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/1600_concurrent_consumers-10       5.028µ ±  8%                       2.574µ ±  5%   -48.80% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/16_concurrent_consumers-10         529.9n ±  1%                      1765.5n ±  2%  +233.18% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/160_concurrent_consumers-10        699.4n ±  1%                      2051.5n ±  8%  +193.34% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/1600_concurrent_consumers-10       2.372µ ±  7%                       2.488µ ±  8%         ~ (p=0.180 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/16_concurrent_consumers-10        813.5n ±  3%                      1748.0n ±  3%  +114.86% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/160_concurrent_consumers-10       1.539µ ±  4%                       2.078µ ± 10%   +35.07% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/1600_concurrent_consumers-10      7.961µ ±  9%                       2.357µ ± 20%   -70.39% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/16_concurrent_consumers-10        513.9n ±  1%                      1746.5n ±  2%  +239.85% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/160_concurrent_consumers-10       1.174µ ±  2%                       1.846µ ±  6%   +57.31% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/1600_concurrent_consumers-10      4.127µ ±  8%                       2.335µ ± 15%   -43.41% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/16_concurrent_consumers-10      1.347µ ±  2%                       1.858µ ±  2%   +37.94% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/160_concurrent_consumers-10     2.344µ ±  2%                       1.972µ ±  5%   -15.87% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/1600_concurrent_consumers-10   11.358µ ± 10%                       2.471µ ± 13%   -78.25% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/16_concurrent_consumers-10      554.2n ±  0%                      1817.0n ±  2%  +227.86% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/160_concurrent_consumers-10     1.901µ ±  2%                       1.915µ ± 12%         ~ (p=1.000 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/1600_concurrent_consumers-10    6.182µ ±  3%                       2.583µ ±  7%   -58.23% (p=0.002 n=6)
geomean                                                                                        1.693µ                             2.062µ         +21.81%

                                                                                            │ original.txt │ step-9-simpler-context-cancellation-with-less-memory.txt │
                                                                                            │     B/op     │               B/op                vs base                │
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/16_concurrent_consumers-10         199.5 ±  1%                        402.0 ±  1%  +101.50% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/160_concurrent_consumers-10        265.0 ±  3%                        467.0 ±  3%   +76.23% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/1600_concurrent_consumers-10      1015.0 ±  3%                        533.0 ±  5%   -47.49% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/16_concurrent_consumers-10         184.5 ±  1%                        402.0 ±  1%  +117.89% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/160_concurrent_consumers-10        225.0 ±  1%                        472.0 ±  5%  +109.78% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/1600_concurrent_consumers-10       536.0 ±  4%                        527.5 ±  6%         ~ (p=0.310 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/16_concurrent_consumers-10        587.0 ±  4%                        391.0 ±  2%   -33.39% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/160_concurrent_consumers-10       796.5 ±  6%                        447.5 ±  7%   -43.82% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/1600_concurrent_consumers-10     1913.0 ±  8%                        500.5 ± 21%   -73.84% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/16_concurrent_consumers-10        100.5 ±  6%                        389.0 ±  2%  +287.06% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/160_concurrent_consumers-10       739.0 ±  3%                        406.0 ±  7%   -45.06% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/1600_concurrent_consumers-10     1234.5 ±  7%                        480.0 ± 13%   -61.12% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/16_concurrent_consumers-10     1944.5 ±  5%                        404.5 ±  1%   -79.20% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/160_concurrent_consumers-10    2170.5 ±  2%                        432.5 ±  6%   -80.07% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/1600_concurrent_consumers-10   3744.0 ± 14%                        519.0 ±  6%   -86.14% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/16_concurrent_consumers-10      88.50 ±  2%                       395.00 ±  3%  +346.33% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/160_concurrent_consumers-10    2102.5 ±  3%                        427.0 ± 12%   -79.69% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/1600_concurrent_consumers-10   2837.0 ±  2%                        545.5 ±  6%   -80.77% (p=0.002 n=6)
geomean                                                                                        673.7                              449.2         -33.31%

                                                                                            │ original.txt │ step-9-simpler-context-cancellation-with-less-memory.txt │
                                                                                            │  allocs/op   │            allocs/op              vs base                │
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/16_concurrent_consumers-10         6.000 ±  0%                        9.000 ±  0%   +50.00% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/160_concurrent_consumers-10        7.000 ±  0%                       11.000 ±  0%   +57.14% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/10_concurrent_producers/1600_concurrent_consumers-10       16.50 ±  9%                        12.00 ±  8%   -27.27% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/16_concurrent_consumers-10         6.000 ±  0%                        9.000 ±  0%   +50.00% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/160_concurrent_consumers-10        6.000 ±  0%                       11.000 ±  9%   +83.33% (p=0.002 n=6)
ConcurrentQueueOperations/1_tenants/25_concurrent_producers/1600_concurrent_consumers-10       10.00 ± 10%                        12.00 ±  8%   +20.00% (p=0.006 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/16_concurrent_consumers-10        6.000 ±  0%                        9.000 ±  0%   +50.00% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/160_concurrent_consumers-10       8.000 ±  0%                       10.000 ± 10%   +25.00% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/10_concurrent_producers/1600_concurrent_consumers-10      22.00 ±  9%                        11.50 ± 22%   -47.73% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/16_concurrent_consumers-10        5.000 ±  0%                        9.000 ±  0%   +80.00% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/160_concurrent_consumers-10       7.000 ±  0%                        9.000 ± 11%   +28.57% (p=0.002 n=6)
ConcurrentQueueOperations/10_tenants/25_concurrent_producers/1600_concurrent_consumers-10      13.50 ±  4%                        10.50 ± 14%   -22.22% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/16_concurrent_consumers-10      9.000 ± 11%                       10.000 ±  0%   +11.11% (p=0.015 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/160_concurrent_consumers-10     12.00 ±  0%                        11.00 ±  9%    -8.33% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/10_concurrent_producers/1600_concurrent_consumers-10    30.50 ±  8%                        12.50 ±  4%   -59.02% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/16_concurrent_consumers-10      5.000 ±  0%                       10.000 ±  0%  +100.00% (p=0.002 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/160_concurrent_consumers-10     11.00 ±  0%                        10.50 ± 14%         ~ (p=0.182 n=6)
ConcurrentQueueOperations/1000_tenants/25_concurrent_producers/1600_concurrent_consumers-10    19.00 ±  5%                        13.00 ±  0%   -31.58% (p=0.002 n=6)
geomean                                                                                        9.540                              10.48          +9.87%

```
</details>

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated
- [n/a] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`