@charleskorn 2023-09-01
@dimitarvdimitrov

> I'm not super familiar with the scheduler, but I think I understand your changes. So this effectively trades the contention on a sync.Cond with contention on a go channel (`availableQueriers`), did I get this right?

Yes, and it also removes the issue where a goroutine per waiting querier is woken up to check for work every time we receive a new query, even if there are none waiting by the time that querier's goroutine takes the lock. 

Now, when a new query arrives, we'll only do an amount of work in proportion to `min(queriers waiting, queries waiting)`. Unless shuffle sharding is enabled, that expression will either be 1 (there was a querier waiting when the query arrived) or 0 (there were no queriers waiting).