@charleskorn 2023-09-01
I've updated the benchmark results in the PR description based on the most recent changes.

There's a regression due to two changes:
* [0498225](https://github.com/grafana/mimir/pull/5880/commits/0498225cc9e123e4085619ebd4c5d619b4fd49c1): necessary for correctness, regression is small
* [5944299](https://github.com/grafana/mimir/pull/5880/commits/5944299f806eccb2a782a395e181730836e522e4): necessary for correctness, regression is larger than I'd like but still OK compared to the previous implementation for large numbers of queriers