@charleskorn 2023-09-01
I think this is good enough to review. There are a couple more performance optimisations I'd like to try do but I'll keep them for a follow-up PR in the interests of keeping this one small.

Ideas I'd like to try:

* pool `enqueueRequest` and `availableQuerier` structs
* store the value of `q.queueLength.WithLabelValues(userID)` on the `userQueue`, to save retrieving it every time we enqueue or dispatch a request
* only trigger dispatching queries after querier operations if they actually resulted in resharding of queriers
* only trigger dispatching queries after enqueuing a request if it's the first request for that tenant, or we resharded queriers because `maxQueriers` changed