@dimitarvdimitrov 2023-09-01
I see you mentioned you want to keep the improvements in a follow-up PR. Is it ok to track the work items in issues instead of TODOs in the code? There are some of long-forgotten TODOs that lay around the repo that no one addresses and was hoping to not contribute to that 😅  

Also the integration tests are failing with some data races in the `scheduler/queue.(*RequestQueue)`. Have you seen those?