@charleskorn 2023-09-04
I've greatly simplified the context cancellation handling in https://github.com/grafana/mimir/pull/5880/files/da9874a811d2fbfc47097a28419103e6f7e67de9..dc1ecf59aba2e3f2c79b8553bc9fff172f9523de. Not only does it make the code far simpler, it also improves performance as we're not spinning up an extra goroutine for each `GetNextRequestForQuerier` call.

I've updated the benchmark results in the PR description to match.